billing_api/requirements.txt
fastapi==0.103.2
orjson==3.9.7
pydantic==2.6.0
uvicorn==0.12.2
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
gunicorn==21.2.0
httptools==0.6.0
backoff==1.10.0
python-dotenv==1.0.0
aiohttp==3.9.3
pydantic-settings
alembic==1.11.1
asyncpg==0.27.0
redis==4.5.5
yookassa==3.1.0
python-jose==3.3.0
sqladmin==0.16.1
asgi-correlation-id==4.3.1
requests
itsdangerous
